requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
